refactor; make migrate_tables a module
